backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
joblib
openpyxl
google-generativeai
markdown
xhtml2pdf


python-dotenv
